Split io::UndirectedGraph into graph:: , and move io:: to io::graph::
