+++ metis/src/io/graph.rs
@@ -2,6 +2,16 @@
 
 use std::str::FromStr;
 
+pub trait FromMetisGraphFormat: Sized {
+    fn from_metis_graph_format_lines(
+        lines: impl Iterator<Item = String>,
+    ) -> Result<Self, GraphFileError>;
+
+    fn from_metis_graph_format_str(input: &str) -> Result<Self, GraphFileError> {
+        Self::from_metis_graph_format_lines(input.trim().lines().map(|line| line.to_string()))
+    }
+}
+
 #[derive(Debug, PartialEq, Clone, thiserror::Error)]
 pub enum LineError {
     #[error("vertex size `s` in manual is missing")]
@@ -54,10 +64,10 @@ pub enum GraphFileError {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq)]
-struct Format {
-    has_vertex_size: bool,
-    has_vertex_weight: bool,
-    has_edge_weight: bool,
+pub struct Format {
+    pub has_vertex_size: bool,
+    pub has_vertex_weight: bool,
+    pub has_edge_weight: bool,
 }
 
 impl FromStr for Format {
@@ -93,14 +103,14 @@ impl Default for Format {
 
 /// Header of METIS Graph file
 #[derive(Debug, Clone, PartialEq)]
-struct Header {
+pub struct Header {
     /// Number of vertices
-    num_vertices: usize,
+    pub num_vertices: usize,
     /// Number of edges
-    num_edges: usize,
-    fmt: Format,
+    pub num_edges: usize,
+    pub fmt: Format,
     /// Number of vertex weights associated with each vertex of the graph
-    num_weights: usize,
+    pub num_weights: usize,
 }
 
 impl FromStr for Header {
@@ -132,22 +142,22 @@ impl FromStr for Header {
 }
 
 #[derive(Debug)]
-struct Line {
+pub struct Line {
     /// `s` in manual
     /// None if Header.has_vertex_size is false
-    vertex_size: Option<i32>,
+    pub vertex_size: Option<i32>,
     /// `w_1`, `w_2`, ... in manual
     /// None if Header.has_vertex_weight is false
-    vertex_weights: Option<Vec<f32>>,
+    pub vertex_weights: Option<Vec<f32>>,
     /// `v1`, ... in manual
-    vertices: Vec<i32>,
+    pub vertices: Vec<i32>,
     /// `e1`, ... in manual
     /// None if Header.has_edge_weight is false
-    edge_weights: Option<Vec<f32>>,
+    pub edge_weights: Option<Vec<f32>>,
 }
 
 impl Line {
-    fn parse(header: &Header, line: &str) -> Result<Self, LineError> {
+    pub fn parse(header: &Header, line: &str) -> Result<Self, LineError> {
         let mut nums = line.trim().split_whitespace();
         let vertex_size = if header.fmt.has_vertex_size {
             let s = nums.next().ok_or(LineError::VertexSizeMissing)?;
@@ -194,54 +204,6 @@ impl Line {
             vertices,
             edge_weights,
         })
-    }
-}
-
-/// uncompressed graph
-#[derive(Debug, Clone)]
-pub struct UndirectedGraph {
-    vertex_size: usize,
-    edges: Vec<(i32, i32)>,
-}
-
-impl UndirectedGraph {
-    fn from_lines(mut lines: impl Iterator<Item = String>) -> Result<Self, GraphFileError> {
-        let header = Header::from_str(
-            &lines
-                .next()
-                .ok_or(GraphFileError::InvalidHeader(HeaderError::Empty))?,
-        )?;
-        let mut edges = Vec::new();
-        for (from_index, line) in lines.enumerate() {
-            let from_index = from_index as i32 + 1;
-            let parsed =
-                Line::parse(&header, &line).map_err(|error| GraphFileError::InvalidLine {
-                    error,
-                    line_position: from_index as usize,
-                })?;
-            for &to_index in &parsed.vertices {
-                if from_index < to_index {
-                    edges.push((from_index, to_index));
-                }
-            }
-        }
-        if edges.len() != header.num_edges {
-            return Err(GraphFileError::EdgeSizeMissmatch {
-                actual: edges.len(),
-                header: header.num_edges,
-            });
-        }
-        Ok(UndirectedGraph {
-            vertex_size: header.num_vertices as usize,
-            edges,
-        })
-    }
-}
-
-impl FromStr for UndirectedGraph {
-    type Err = GraphFileError;
-    fn from_str(input: &str) -> Result<Self, Self::Err> {
-        UndirectedGraph::from_lines(input.trim().lines().map(|line| line.to_string()))
     }
 }
 
@@ -367,69 +329,5 @@ mod tests {
                 }
             );
         }
-    }
-
-    #[test]
-    fn unweighted_graph() {
-        // graph in Figure 2.(a)
-        let input = r#"
-            7 11
-            5 3 2
-            1 3 4
-            5 4 2 1
-            2 3 6 7
-            1 3 6
-            5 4 7
-            6 4
-        "#;
-        let _graph = UndirectedGraph::from_str(input).unwrap();
-    }
-
-    #[test]
-    fn weighted_graph_weights_on_edges() {
-        // graph in Figure 2.(b)
-        let input = r#"
-            7 11 001
-            5 1 3 2 2 1
-            1 1 3 2 4 1
-            5 3 4 2 2 2 1 2
-            2 1 3 2 6 2 7 5
-            1 1 3 3 6 2
-            5 2 4 2 7 6
-            6 6 4 5
-        "#;
-        let _graph = UndirectedGraph::from_str(input).unwrap();
-    }
-
-    #[test]
-    fn weighted_graph_weights_both_on_vertices_and_edges() {
-        // graph in Figure 2.(c)
-        let input = r#"
-            7 11 011
-            4 5 1 3 2 2 1
-            2 1 1 3 2 4 1
-            5 5 3 4 2 2 2 1 2
-            3 2 1 3 2 6 2 7 5
-            1 1 1 3 3 6 2
-            6 5 2 4 2 7 6
-            2 6 6 4 5
-        "#;
-        let _graph = UndirectedGraph::from_str(input).unwrap();
-    }
-
-    #[test]
-    fn multi_constraint_graph() {
-        // graph in Figure 2.(d)
-        let input = r#"
-            7 11 010 3
-            1 2 0 5 3 2
-            0 2 2 1 3 4
-            4 1 1 5 4 2 1
-            2 2 3 2 3 6 7
-            1 1 1 1 3 6
-            2 2 1 5 4 7
-            1 2 1 6 4
-        "#;
-        let _graph = UndirectedGraph::from_str(input).unwrap();
     }
 }
