+++ metis/src/lib.rs
@@ -1,4 +1,5 @@
 //! Rust binding of [METIS](http://glaros.dtc.umn.edu/gkhome/metis/metis/overview)
 
 pub mod error;
+pub mod graph;
 pub mod io;
